implement ID_S3_EX2

+++ misc/objdet_tools.py
@@ -262,7 +262,7 @@ def project_detections_into_bev(bev_map, detections, configs, color=[]):
         cv2.polylines(bev_map, [corners_int], True, color, 2)
 
         # draw colored line to identify object front
-        corners_int = bev_corners.reshape(-1, 2)
+        corners_int = bev_corners.reshape(-1, 2).astype(int)
         cv2.line(
             bev_map,
             (corners_int[0, 0], corners_int[0, 1]),
